render voice names for Cloud voices
